requirements.txt
aerich==0.5.8
aiofiles==0.8.0
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.1
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
ddlparse==1.10.0
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-admin==1.0.3
fastapi-mail==1.0.2
h11==0.12.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.19.0
idna==3.3
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.4
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
pypika-tortoise==0.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
redis==4.0.2
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
tortoise-orm==0.17.8
typing_extensions==4.0.0
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
